refactor recipe delete
