+++ src/frontend/components/recipes/RecipeDetails.vue
@@ -1,7 +1,12 @@
 <script setup lang="ts">
 import { convertNewlinesToBr, formatDate } from '@/utils/formatUtils'
+import { ref } from 'vue'
+import { doErrorHandling } from '@/utils/generalUtils'
+import { deleteRecipe } from '@/services/RecipeService'
+import { snackbarStore } from '@/stores/snackbarStore'
+import router from '@/router'
 
-defineProps<{
+const { recipeId } = defineProps<{
   recipeId: number;
   instructions: string;
   notes: string;
@@ -10,33 +15,24 @@ defineProps<{
   showEditMode: boolean;
 }>();
 
+const showDeleteConfirmation = ref<boolean>(false)
+
+const doDeleteRecipe = async () => {
+  await doErrorHandling(async () => {
+    await deleteRecipe(recipeId)
+    snackbarStore.showSuccessMessage("Recipe deleted")
+    await router.push("/recipes")
+  }, "deleting recipe")
+}
 </script>
 
 <template>
-<!--  <v-textarea-->
-<!--    v-if="showEditMode"-->
-<!--    v-model="recipe.instructions"-->
-<!--    label="Instructions"-->
-<!--    placeholder="How to actually make this"-->
-<!--    :rows="3"-->
-<!--    auto-grow-->
-<!--    outlined-->
-<!--  />-->
   <div>
     <h3>Instructions:</h3>
     <p v-if="instructions == ''"><em>No instructions</em></p>
     <p v-else v-html="convertNewlinesToBr(instructions)"/>
   </div>
 
-<!--  <v-textarea-->
-<!--    v-if="showEditMode"-->
-<!--    v-model="recipe.notes"-->
-<!--    label="Notes"-->
-<!--    placeholder="Things to remember about this recipe"-->
-<!--    :rows="2"-->
-<!--    auto-grow-->
-<!--    outlined-->
-<!--  />-->
   <div>
     <hr style="width: 100%;"/>
     <h3>Notes:</h3>
@@ -48,7 +44,24 @@ defineProps<{
 
   <p>Created {{formatDate(created_at)}}</p>
   <p>Last edited {{formatDate(updated_at)}}</p>
-<!--  <v-btn v-if="showEditMode" color="red" @click="true">Delete Recipe</v-btn>-->
+  <v-btn v-if="showEditMode || true" color="red" @click="showDeleteConfirmation = true">Delete Recipe</v-btn>
+
+
+  <v-dialog
+    v-model="showDeleteConfirmation"
+    width="auto"
+    min-width="300px">
+    <v-card title="Are you sure?">
+      <template v-slot:text>
+        <p>Do you want to delete this recipe?</p>
+        <p><em>You can not undo this</em></p>
+      </template>
+      <template v-slot:actions>
+        <v-btn @click="doDeleteRecipe" color="red">Delete</v-btn>
+        <v-btn @click="showDeleteConfirmation = false">Cancel</v-btn>
+      </template>
+    </v-card>
+  </v-dialog>
 </template>
 
 <style scoped>
